Realizado Testing del proyecto
Se han creado una clase para Triangulo con 2 casos de prueba y una clase control_TrianguloTest para la clase control_Triangulo con 12 casos de prueba.

+++ Triangulo/src/main/java/ISO_C01_P2/Triangulo/Triangulo.java
@@ -1,9 +1,9 @@
 package ISO_C01_P2.Triangulo;
 
 public class Triangulo {
-	public int lado1, lado2,lado3;
+	public double lado1, lado2,lado3;
 	public double angulo1, angulo2, angulo3;
-	public Triangulo(int lado1, int lado2, int lado3, double angulo1, double angulo2, double angulo3) {
+	public Triangulo(double lado1, double lado2, double lado3, double angulo1, double angulo2, double angulo3) {
 		super();
 		this.lado1 = lado1;
 		this.lado2 = lado2;
@@ -12,22 +12,22 @@ public class Triangulo {
 		this.angulo2 = angulo2;
 		this.angulo3 = angulo3;
 	}
-	public int getLado1() {
+	public double getLado1() {
 		return lado1;
 	}
-	public void setLado1(int lado1) {
+	public void setLado1(double lado1) {
 		this.lado1 = lado1;
 	}
-	public int getLado2() {
+	public double getLado2() {
 		return lado2;
 	}
-	public void setLado2(int lado2) {
+	public void setLado2(double lado2) {
 		this.lado2 = lado2;
 	}
-	public int getLado3() {
+	public double getLado3() {
 		return lado3;
 	}
-	public void setLado3(int lado3) {
+	public void setLado3(double lado3) {
 		this.lado3 = lado3;
 	}
 	public double getAngulo1() {
